backend/requirements.txt
fastapi
sqlalchemy
uvicorn
pydantic
bcrypt
python-dotenv
psycopg2-binary
python-jose
